강의 외
- 전략패턴과 템플릿콜백패턴 관련 클래스 분리를 위해 패키지 이동
