+++ src/test/java/hello/advanced/trace/strategy/code/strategy/Strategy.java
@@ -1,4 +1,4 @@
-package hello.advanced.trace.strategy.code;
+package hello.advanced.trace.strategy.code.strategy;
 
 // 변하는 알고리즘 역할
 public interface Strategy {
